requirements.txt
attrs==23.2.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.8
filelock==3.15.1
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fsspec==2024.6.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
imageio==2.34.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.0
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMatting==1.1.12
pyparsing==3.1.2
pyreadline3==3.4.1
PyYAML==6.0.1
referencing==0.35.1
rembg==2.0.57
requests==2.32.3
rpds-py==0.18.1
rsa==4.9
scikit-image==0.23.2
scipy==1.13.1
super-image==0.1.7
sympy==1.12.1
tbb==2021.12.0
tifffile==2024.5.22
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3